requirements.txt
mysql-connector-python==8.4.0
aiosmtplib==3.0.1
bilibili-api-dev==16.2.0b1